requirements.txt
    # Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client
requests>=2.31.0

# Data Validation
pydantic>=2.5.0
pydantic[email]>=2.5.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Environment Management
python-dotenv>=1.0.0

# Development Dependencies (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
# Testing
fastapi[all]>=0.104.0

# Background Job Scheduling
apscheduler>=3.10.4